Cherry-pick Move sequence_unpad from x86 to host, test=develop
