+++ lite/backends/host/CMakeLists.txt
@@ -1,3 +1,5 @@
 lite_cc_library(target_wrapper_host SRCS target_wrapper.cc)
+
+add_subdirectory(math)
