+++ lite/backends/host/math/CMakeLists.txt
@@ -1,0 +1,1 @@
+lite_cc_library(sequence_padding SRCS sequence_padding.cc DEPS ${lite_kernel_deps})
